Replace main scenes
